+++ source/renderer/app/containers/wallet/dialogs/WalletConnectDialogContainer.js
@@ -26,6 +26,8 @@ export default class WalletConnectDialogContainer extends Component<Props> {
     const { hardwareWallets, wallets, app } = stores;
     const { hwDeviceStatus, transportDevice } = hardwareWallets;
     const { createHardwareWalletRequest } = wallets;
+    const { isLinux } = app.environment;
+
     return (
       <WalletConnectDialog
         isSubmitting={createHardwareWalletRequest.isExecuting}
@@ -33,6 +35,7 @@ export default class WalletConnectDialogContainer extends Component<Props> {
         onClose={this.onClose}
         hwDeviceStatus={hwDeviceStatus}
         transportDevice={transportDevice}
+        isLinux={isLinux}
         onExternalLinkClick={app.openExternalLink}
       />
     );