+++ storybook/stories/wallets/hardwareWallets/HardwareWallets.stories.js
@@ -1,6 +1,7 @@
 // @flow
 import React from 'react';
 import { storiesOf } from '@storybook/react';
+import { boolean } from '@storybook/addon-knobs';
 import { action } from '@storybook/addon-actions';
 import WalletConnectDialog from '../../../../source/renderer/app/components/wallet/WalletConnectDialog';
 import HardwareWalletsWrapper from '../_utils/HardwareWalletsWrapper';
@@ -39,6 +40,7 @@ storiesOf('Wallets|Hardware Wallets', module)
 
   .add('Hardware wallet connect Ledger step 1', () => (
     <WalletConnectDialog
+      isLinux={boolean('isLinux', false)}
       transportDevice={ledgerDevice}
       hwDeviceStatus={HwDeviceStatuses.CONNECTING}
       onClose={action('onClose')}
@@ -50,6 +52,7 @@ storiesOf('Wallets|Hardware Wallets', module)
 
   .add('Hardware wallet connect Ledger step 2', () => (
     <WalletConnectDialog
+      isLinux={boolean('isLinux', false)}
       transportDevice={ledgerDevice}
       hwDeviceStatus={HwDeviceStatuses.LAUNCHING_CARDANO_APP}
       onClose={action('onClose')}
@@ -61,6 +64,7 @@ storiesOf('Wallets|Hardware Wallets', module)
 
   .add('Hardware wallet connect Ledger step 3', () => (
     <WalletConnectDialog
+      isLinux={boolean('isLinux', false)}
       transportDevice={ledgerDevice}
       hwDeviceStatus={HwDeviceStatuses.EXPORTING_PUBLIC_KEY}
       onClose={action('onClose')}
@@ -72,6 +76,7 @@ storiesOf('Wallets|Hardware Wallets', module)
 
   .add('Hardware wallet connect Ledger step 4', () => (
     <WalletConnectDialog
+      isLinux={boolean('isLinux', false)}
       transportDevice={ledgerDevice}
       hwDeviceStatus={HwDeviceStatuses.EXPORTING_PUBLIC_KEY_FAILED}
       onClose={action('onClose')}
@@ -83,6 +88,7 @@ storiesOf('Wallets|Hardware Wallets', module)
 
   .add('Hardware wallet connect Ledger step 5', () => (
     <WalletConnectDialog
+      isLinux={boolean('isLinux', false)}
       transportDevice={ledgerDevice}
       hwDeviceStatus={HwDeviceStatuses.READY}
       onClose={action('onClose')}
@@ -94,6 +100,7 @@ storiesOf('Wallets|Hardware Wallets', module)
 
   .add('Hardware wallet connect Trezor step 1', () => (
     <WalletConnectDialog
+      isLinux={boolean('isLinux', false)}
       transportDevice={trezorDevice}
       hwDeviceStatus={HwDeviceStatuses.CONNECTING}
       onClose={action('onClose')}
@@ -105,6 +112,7 @@ storiesOf('Wallets|Hardware Wallets', module)
 
   .add('Hardware wallet connect Trezor step 2', () => (
     <WalletConnectDialog
+      isLinux={boolean('isLinux', false)}
       transportDevice={trezorDevice}
       hwDeviceStatus={HwDeviceStatuses.LAUNCHING_CARDANO_APP}
       onClose={action('onClose')}
@@ -116,6 +124,7 @@ storiesOf('Wallets|Hardware Wallets', module)
 
   .add('Hardware wallet connect Trezor step 3', () => (
     <WalletConnectDialog
+      isLinux={boolean('isLinux', false)}
       transportDevice={trezorDevice}
       hwDeviceStatus={HwDeviceStatuses.EXPORTING_PUBLIC_KEY}
       onClose={action('onClose')}
@@ -127,6 +136,7 @@ storiesOf('Wallets|Hardware Wallets', module)
 
   .add('Hardware wallet connect Trezor step 4', () => (
     <WalletConnectDialog
+      isLinux={boolean('isLinux', false)}
       transportDevice={trezorDevice}
       hwDeviceStatus={HwDeviceStatuses.EXPORTING_PUBLIC_KEY_FAILED}
       onClose={action('onClose')}
@@ -138,6 +148,7 @@ storiesOf('Wallets|Hardware Wallets', module)
 
   .add('Hardware wallet connect Trezor step 5', () => (
     <WalletConnectDialog
+      isLinux={boolean('isLinux', false)}
       transportDevice={trezorDevice}
       hwDeviceStatus={HwDeviceStatuses.READY}
       onClose={action('onClose')}