+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changelog
 
 ### Chores
 
+- Added "Fix connection Issues" link to the hardware wallet pairing dialog ([PR 2336](https://github.com/input-output-hk/daedalus/pull/2336))
 - Fixed error thrown when closing delegation wizard while transaction fees are being calculated ([PR 2330](https://github.com/input-output-hk/daedalus/pull/2330))
 - Fixed number format for syncing percentage and stake pools count ([PR 2313](https://github.com/input-output-hk/daedalus/pull/2313))
 - Updated `cardano-wallet` to version `2021-01-28` and `cardano-node` to version `1.25.1` ([PR 2270](https://github.com/input-output-hk/daedalus/pull/2270))